Fix uninstall not working after v1.0

+++ uninstall.sh
@@ -2,7 +2,8 @@
 set -euo pipefail
 
 
-DESTINATION="$HOME/.mandarin-duck"
+: "${DESTINATION:="$HOME/.mandarin-duck"}"
+FOUND_VERSION=$(jq --raw-output ".version" "$DESTINATION/mandarin-duck.cfg")
 REPOS=$(jq --raw-output ".projects | keys | .[]" "$DESTINATION/mandarin-duck.cfg")
 BUILDKITE_API_TOKEN=$(jq --raw-output ".buildkite_api_token" "$DESTINATION/mandarin-duck.cfg")
 DESTROY_COUNT=$(wc -l <<<"$REPOS" | tr -d " ")
@@ -20,9 +21,17 @@ fi
 
 
 echo "Uninstalling mandarin-duck..."
-while read -r REPO;do
-    sed -i "/# mandarin-duck v[.0-9]*/d" "$REPO/hooks/post-receive"
+while read -r REPO; do
+    if [[ $FOUND_VERSION == "1.0" ]]; then
+        sed -i "/# mandarin-duck v[.0-9]*/d" "$REPO/hooks/post-receive"
+    else
+        unlink "$REPO/hooks/post-receive"
+    fi
 done <<<"$REPOS"; 
-curl --fail --silent -X DELETE "https://api.buildkite.com/v2/access-token" \
-    -H "Authorization: Bearer $BUILDKITE_API_TOKEN"
+if [[ $SKIP_BUILDKITE_API_CALLS_BECAUSE_TESTING != "true" ]]; then
+    curl --fail --silent -X DELETE "https://api.buildkite.com/v2/access-token" \
+        -H "Authorization: Bearer $BUILDKITE_API_TOKEN"
+else
+    echo "Skipping Buildkite API call..."
+fi
 rm -rf "$DESTINATION"
